requirements.txt
colorama==0.4.3
Flask==1.1.1
numpy==1.16.1
requests==2.20.0
plotly==4.5.4
pandas==0.25.3
psycopg2-binary==2.8.5
scipy==1.2.2
Werkzeug==0.16.0
termcolor==1.1.0
ccxt==1.32.3
empyrical==0.5.3
flask_httpauth==3.3.0
nest_asyncio==1.3.2
pyfiglet==0.8.post1
slackclient==2.5.0
ta==0.5.20
pylint
pytest-cov
robin_stocks
cython-indicators==0.0.6
matplotlib
sklearn
python-telegram-bot
